add extra command
